Refactor examples/snapshot
