+++ examples/snapshot/Clock.js
