+++ examples/snapshot/__tests__/clock.test.js
@@ -3,7 +3,7 @@
 'use strict';
 
 import React from 'react';
-import Clock from '../Clock.react';
+import Clock from '../Clock';
 import renderer from 'react-test-renderer';
 
 jest.useFakeTimers();
